polymorphism broken in broker so everything is bound to uist, removed dividends
