+++ crates/alator/Cargo.toml
@@ -20,7 +20,7 @@ async-trait = "0.1.73"
 tokio = { version = "1.32.0", features = ["full"] }
 futures = "0.3.28"
 alator-clock = { path = "../alator-clock" }
-alator-exchange = { path = "../alator-exchange" }
+rotala = { path = "../rotala" }
 
 [dev-dependencies]
 reqwest = { version = "0.11.11", features=["blocking"] }
