+++ crates/alator/src/simcontext/builder.rs
@@ -1,6 +1,6 @@
 use crate::strategy::{History, Strategy};
 use crate::types::CashValue;
-use alator_clock::Clock;
+use rotala::clock::Clock;
 
 use super::SimContext;
