+++ crates/alator/src/strategy/staticweight.rs
@@ -1,18 +1,15 @@
 use log::info;
+use rotala::exchange::uist::UistTrade;
 
-use crate::broker::single::SingleBroker;
-use crate::broker::{
-    BacktestBroker, BrokerCalculations, BrokerCashEvent, DividendPayment, EventLog, ReceivesOrders,
-    Trade,
-};
+use crate::broker::uist::{UistBroker, UistBrokerCashEvent};
 use crate::schedule::{DefaultTradingSchedule, TradingSchedule};
 use crate::strategy::{Audit, History, Strategy, StrategyEvent, TransferFrom, TransferTo};
 use crate::types::{CashValue, PortfolioAllocation, StrategySnapshot};
-use alator_clock::Clock;
+use rotala::clock::Clock;
 
 pub struct StaticWeightStrategyBuilder {
     //If missing either field, we cannot run this strategy
-    brkr: Option<SingleBroker>,
+    brkr: Option<UistBroker>,
     weights: Option<PortfolioAllocation>,
     clock: Option<Clock>,
 }
@@ -39,7 +36,7 @@ impl StaticWeightStrategyBuilder {
         self
     }
 
-    pub fn with_brkr(&mut self, brkr: SingleBroker) -> &mut Self {
+    pub fn with_brkr(&mut self, brkr: UistBroker) -> &mut Self {
         self.brkr = Some(brkr);
         self
     }
@@ -67,7 +64,7 @@ impl Default for StaticWeightStrategyBuilder {
 ///Basic implementation of an investment strategy which takes a set of fixed-weight allocations and
 ///rebalances over time towards those weights.
 pub struct StaticWeightStrategy {
-    brkr: SingleBroker,
+    brkr: UistBroker,
     target_weights: PortfolioAllocation,
     net_cash_flow: CashValue,
     clock: Clock,
@@ -92,9 +89,8 @@ impl Strategy for StaticWeightStrategy {
     fn init(&mut self, initital_cash: &f64) {
         self.deposit_cash(initital_cash);
         if DefaultTradingSchedule::should_trade(&self.clock.now()) {
-            let orders = BrokerCalculations::diff_brkr_against_target_weights(
-                &self.target_weights,
-                &mut self.brkr,
+            let orders = self.brkr.diff_brkr_against_target_weights(
+                &self.target_weights
             );
             if !orders.is_empty() {
                 self.brkr.send_orders(&orders);
@@ -106,9 +102,8 @@ impl Strategy for StaticWeightStrategy {
         self.brkr.check();
         let now = self.clock.now();
         if DefaultTradingSchedule::should_trade(&now) {
-            let orders = BrokerCalculations::diff_brkr_against_target_weights(
-                &self.target_weights,
-                &mut self.brkr,
+            let orders = self.brkr.diff_brkr_against_target_weights(
+                &self.target_weights
             );
             if !orders.is_empty() {
                 self.brkr.send_orders(&orders);
@@ -130,7 +125,7 @@ impl TransferTo for StaticWeightStrategy {
 
 impl TransferFrom for StaticWeightStrategy {
     fn withdraw_cash(&mut self, cash: &f64) -> StrategyEvent {
-        if let BrokerCashEvent::WithdrawSuccess(withdrawn) = self.brkr.withdraw_cash(cash) {
+        if let UistBrokerCashEvent::WithdrawSuccess(withdrawn) = self.brkr.withdraw_cash(cash) {
             info!("STRATEGY: Succesfully withdrew {:?} from strategy", cash);
             self.net_cash_flow = CashValue::from(*self.net_cash_flow - *withdrawn);
             return StrategyEvent::WithdrawSuccess(CashValue::from(*cash));
@@ -140,10 +135,10 @@ impl TransferFrom for StaticWeightStrategy {
     }
 
     fn withdraw_cash_with_liquidation(&mut self, cash: &f64) -> StrategyEvent {
-        if let BrokerCashEvent::WithdrawSuccess(withdrawn) =
+        if let UistBrokerCashEvent::WithdrawSuccess(withdrawn) =
             //No logging here because the implementation is fully logged due to the greater
             //complexity of this task vs standard withdraw
-            BrokerCalculations::withdraw_cash_with_liquidation(cash, &mut self.brkr)
+            self.brkr.withdraw_cash_with_liquidation(cash)
         {
             self.net_cash_flow = CashValue::from(*self.net_cash_flow - *withdrawn);
             return StrategyEvent::WithdrawSuccess(CashValue::from(*cash));
@@ -153,12 +148,8 @@ impl TransferFrom for StaticWeightStrategy {
 }
 
 impl Audit for StaticWeightStrategy {
-    fn trades_between(&self, start: &i64, end: &i64) -> Vec<Trade> {
+    fn trades_between(&self, start: &i64, end: &i64) -> Vec<UistTrade> {
         self.brkr.trades_between(start, end)
-    }
-
-    fn dividends_between(&self, start: &i64, end: &i64) -> Vec<DividendPayment> {
-        self.brkr.dividends_between(start, end)
     }
 }
