+++ crates/alator/tests/staticweight_test.rs
@@ -1,35 +1,35 @@
-use alator_clock::{Clock, ClockBuilder, Frequency};
-use alator_exchange::input::DefaultPriceSource;
-use alator_exchange::SyncExchangeImpl;
+use alator::broker::types::BrokerCost;
+use alator::broker::uist::UistBrokerBuilder;
+use rotala::clock::Frequency;
 use rand::distributions::{Distribution, Uniform};
 use rand::thread_rng;
 
-use alator::broker::single::SingleBrokerBuilder;
-use alator::broker::BrokerCost;
 use alator::simcontext::SimContextBuilder;
 use alator::strategy::staticweight::StaticWeightStrategyBuilder;
 use alator::types::{CashValue, PortfolioAllocation};
+use rotala::exchange::uist::Uist;
+use rotala::input::penelope::PenelopeBuilder;
 
-fn build_data(clock: Clock) -> DefaultPriceSource {
+fn build_data(length: i64) -> PenelopeBuilder {
     let price_dist = Uniform::new(90.0, 100.0);
     let mut rng = thread_rng();
 
-    let mut price_source = DefaultPriceSource::new();
-    for date in clock.peek() {
-        price_source.add_quotes(
+    let mut source = PenelopeBuilder::new();
+    for date in 1..length+1 {
+        source.add_quote(
             price_dist.sample(&mut rng),
             price_dist.sample(&mut rng),
-            *date,
+            date,
             "ABC",
         );
-        price_source.add_quotes(
+        source.add_quote(
             price_dist.sample(&mut rng),
             price_dist.sample(&mut rng),
-            *date,
+            date,
             "BCD",
         );
     }
-    price_source
+    source
 }
 
 #[test]
@@ -37,26 +37,23 @@ fn staticweight_integration_test() {
     env_logger::init();
     let initial_cash: CashValue = 100_000.0.into();
     let length_in_days: i64 = 1000;
-    let start_date: i64 = 1609750800; //Date - 4/1/21 9:00:0000
-    let clock = ClockBuilder::with_length_in_days(start_date, length_in_days)
-        .with_frequency(&Frequency::Daily)
-        .build();
 
-    let price_source = build_data(clock.clone());
+    let price_source_builder = build_data(1000);
+    let (source, clock) = price_source_builder.build_with_frequency(rotala::clock::Frequency::Second);
 
     let mut weights: PortfolioAllocation = PortfolioAllocation::new();
     weights.insert("ABC", 0.5);
     weights.insert("BCD", 0.5);
 
-    let exchange = SyncExchangeImpl::new(clock.clone(), price_source);
+    let exchange = Uist::new(clock.clone(), source);
 
-    let simbrkr = SingleBrokerBuilder::new()
+    let brkr = UistBrokerBuilder::new()
         .with_trade_costs(vec![BrokerCost::Flat(1.0.into())])
         .with_exchange(exchange)
         .build();
 
     let strat = StaticWeightStrategyBuilder::new()
-        .with_brkr(simbrkr)
+        .with_brkr(brkr)
         .with_weights(weights)
         .with_clock(clock.clone())
         .default();
